+++ fink_filters/ztf/filter_example/filter.py
@@ -20,11 +20,13 @@ from fink_filters.tester import spark_unit_tests
 import pandas as pd
 
 
-def example_filter_(magpsf) -> pd.Series:
-    """Return alerts with difference magnitude above 18
+def example_filter_(cdsxmatch, magpsf) -> pd.Series:
+    """Return alerts with counterpart in the SIMBAD database with difference magnitude above 20
 
     Parameters
     ----------
+    cdsxmatch: Pandas series
+        Column containing the SIMBAD cross-match values
     magpsf: Pandas series
         Column containing difference magnitudes
 
@@ -36,32 +38,41 @@ def example_filter_(magpsf) -> pd.Series:
 
     Examples
     --------
-    >>> pdf = pd.read_parquet('datatest/regular/')
-    >>> classification = example_filter_(pdf['candidate'].apply(lambda x: x['magpsf']))
+    >>> pdf = pd.read_parquet('datatest/magnetic_cvs/')
+    >>> classification = example_filter_(pdf['cdsxmatch'], pdf['candidate'].apply(lambda x: x['magpsf']))
     >>> nalerts = len(pdf[classification]['objectId'])
     >>> print(nalerts)
-    169
+    1
 
     >>> pdf[classification].groupby('cdsxmatch').count().sort_values('objectId', ascending=False)['objectId'].head()
     cdsxmatch
-    LongPeriodV*    16
-    EclBin           8
-    Unknown          8
-    delSctV*         7
-    Mira             7
+    QSO    1
     Name: objectId, dtype: int64
     """
-    mask_mag = magpsf <= 18
+    myfilter = ~cdsxmatch.isin(["Unknown", "Transient"])
 
-    return mask_mag
+    # mask all kind of failures
+    mask = cdsxmatch.apply(lambda x: x.startswith("Fail"))
+    myfilter[mask] = False
+
+    # Remove static objects -- https://github.com/astrolabsoftware/fink-filters/issues/120
+    mask_gal = cdsxmatch.apply(lambda x: x.startswith("Galaxy"))
+    myfilter[mask_gal] = False
+
+    mask_mag = magpsf <= 20.5
+    myfilter[mask_mag] = False
+
+    return myfilter
 
 
 @pandas_udf(BooleanType())
-def example_filter(magpsf: pd.Series) -> pd.Series:
+def example_filter(cdsxmatch: pd.Series, magpsf: pd.Series) -> pd.Series:
     """Pandas UDF version of example_filter_ for Spark
 
     Parameters
     ----------
+    cdsxmatch: Spark DataFrame Column
+        Column containing the SIMBAD cross-match values
     magpsf: Spark DataFrame Column
         Column containing the difference magnitudes
 
@@ -74,13 +85,13 @@ def example_filter(magpsf: pd.Series) -> pd.Series:
     Examples
     --------
     >>> from fink_utils.spark.utils import apply_user_defined_filter
-    >>> df = spark.read.format('parquet').load('datatest/regular/')
+    >>> df = spark.read.format('parquet').load('datatest/magnetic_cvs/')
     >>> f = 'fink_filters.ztf.filter_example.filter.example_filter'
     >>> df = apply_user_defined_filter(df, f)
     >>> print(df.count())
-    169
+    1
     """
-    f_simbad = example_filter_(magpsf)
+    f_simbad = example_filter_(cdsxmatch, magpsf)
 
     return f_simbad
